esboço de cadastro da CC falta melhorar validações

+++ src/conta.java
@@ -1,13 +1,13 @@
 public abstract class conta {
 
-    double numeroConta;
+    int numeroConta;
     String nomeDoTitular;
     protected double saldoAtual;
 
     public conta() {
     }
 
-    public conta(double numeroConta, String nomeDoTitular, double saldoAtual) {
+    public conta(int numeroConta, String nomeDoTitular, double saldoAtual) {
         this.numeroConta = numeroConta;
         this.nomeDoTitular = nomeDoTitular;
         this.saldoAtual = saldoAtual;
