+++ src/contaCorrente.java
@@ -3,7 +3,7 @@ public class contaCorrente extends conta {
     public contaCorrente() {
     }
 
-    public contaCorrente(double numeroConta, String nomeDoTitular, double saldoAtual) {
+    public contaCorrente(int numeroConta, String nomeDoTitular, double saldoAtual) {
         super(numeroConta, nomeDoTitular, saldoAtual);
     }
